src/diploma/requirements.txt
activity==0.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.7.2
Markdown==3.3.4
Pillow==8.1.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.1
